Add files via upload
This is an android application. It is based on tourist guide application which is useful for tourist exploring Pune,India. This app shows all the beautiful places of Pune and also has an option of planning the trip in which you add the places you want to go and this app will plan the the most efficient route for you based on the places you have entered which will save time and also give you the taste of Pune.